+++ src/mapper.js
@@ -1,5 +1,5 @@
 //prettier-ignore
-import { isInteger, isNull, last, deepClone } from "./helpers"
+import { isInteger, isObject, isNull, last, deepClone } from "./helpers"
 
 export const getPath = (path, obj) => {
     const parts = path.split(".")
@@ -19,6 +19,8 @@ export const setPath = (path, value, obj) => {
     const partsMinusLast = parts.slice(0, -1)
     let current = newObj
     partsMinusLast.forEach((part, index) => {
+        checkType(current, part)
+
         const nextPart = parts[index + 1]
         const mustBeAnArray = isInteger(nextPart)
 
@@ -28,7 +30,20 @@ export const setPath = (path, value, obj) => {
         }
         current = current[part]
     })
+
+    checkType(current, last(parts))
     current[last(parts)] = value
 
     return newObj
+}
+
+const checkType = (objOrArray, part) => {
+    const error =
+        (Array.isArray(objOrArray) && !isInteger(part)) ||
+        (isObject(objOrArray) && isInteger(part))
+
+    if (error)
+        throw new Error(
+            `Type mismatch between target type '${typeof objOrArray}' and property '${part}'`,
+        )
 }